Refactor Payment Service: Migrate to BigDecimal for monetary values, restructure event publishing, enhance Gradle dependencies, and update OpenAPI settings.

+++ code/payment-service/src/main/java/vn/id/thongdanghoang/sep/payment_service/PaymentInitiated.java
@@ -1,8 +1,10 @@
 package vn.id.thongdanghoang.sep.payment_service;
 
+import java.math.BigDecimal;
+
 public record PaymentInitiated(
         String transactionId,
         String userId,
-        double originalAmount,
+        BigDecimal originalAmount,
         String voucherCode
 ) {}
